version.conf: bump version number to 376.49_7
This release updates openssl to version 1.0.0r to
in response to the security advisory below.

Description: OpenSSL has issued a security advisory and released updated
versions of SSL containing 12 security fixes.  The vast majority of the
fixes that were included in the advisory were for denial-of-service
issues.
Reference:  https://www.openssl.org/news/secadv_20150319.txt